+++ cli/destroy.ts
@@ -6,9 +6,9 @@ import { deleteInstance } from "../aws/EC2/deleteInstance";
 import { getRunningInstanceIdsByName } from "./getRunningInstanceIdsByName";
 import { getInstanceIdsByPublisher } from "./getInstanceIdsByPublisher";
 import { runWithSpinner } from "./spinner";
-import { getRegion } from "./getRegion";
+import { getRegion } from "./promptUserForRegion";
 import chalk from "chalk";
-import { fetchAllRegions } from "./fetchAllRegions";
+import { fetchAllRegions } from "./getAllEC2Regions";
 
 const deleteAllBrokerInstances = async (regions: string[]) => {
   for (const region of regions) {
