+++ cli/getAllEC2Regions.ts
