+++ cli/promptUserForRegion.ts
